Use project name as basename for interfaces
As noted by Dirk Thomas in ros2/rosidl#441, libraries for generated
interfaces must use ${PROJECT_NAME} as their basename.  This PR
modifies rtabmap_ros so that it does so.

It also tweaks a couple of rosdep keys because nav2_common was used in
some launch files but not included, and message_runtime does not exist
in ROS 2.

Fixes #397